backend/requirements.txt
asgiref==3.7.2
Django==5.0
django-cors-headers==4.3.1
djangorestframework==3.14.0
ifcfg==0.24
psutil==5.9.6
pytz==2023.3.post1
speedtest-cli==2.1.3
sqlparse==0.4.4
typing_extensions==4.9.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
idna==3.3
plotly==5.8.0
requests==2.27.1
tenacity==8.0.1
urllib3==1.26.9

